backend/requirements.txt
fastapi==0.115.6
matplotlib==3.10.0
opencv-python==4.10.0.84
pandas==2.2.3
Pillow==11.1.0
protobuf==5.29.3
psycopg2==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.0
seaborn==0.13.2
setuptools==75.6.0
starlette==0.41.3
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.1
ultralytics==8.3.59
uvicorn==0.34.0
timm==0.4.9
scikit-learn==1.2.2
opencv-contrib-python==4.10.0.84